requirements.txt
Django==1.10.5
django-bootstrap3==7.0.1
django-filter==0.13.0
django-guardian==1.4.4
django-mptt==0.8.7
djangorestframework==3.5.3
